health接口

+++ src/main/java/simplesoul/autumnboot/rest/common/logging/ErrorCounter.java
@@ -1,6 +1,6 @@
 package simplesoul.autumnboot.rest.common.logging;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * 错误日志计数器
@@ -14,12 +14,12 @@ public final class ErrorCounter {
      * {@link LoggerInstrument}向{@link ch.qos.logback.classic.Logger#error(String)}等写入错误日志的方法
      * 插入了调用{@link ErrorCounter#increment()}的逻辑
      */
-    private static final AtomicLong ERRORS = new AtomicLong(0);
+    private static final AtomicInteger ERRORS = new AtomicInteger(0);
 
     /**
      * 获取通过{@link ch.qos.logback.classic.Logger}打印的错误日志数量
      */
-    public static long getErrors() {
+    public static int getErrors() {
         synchronized (ERRORS) {
             return ERRORS.get();
         }
