+++ src/main/java/simplesoul/autumnboot/rest/common/logging/WarningCounter.java
@@ -1,6 +1,6 @@
 package simplesoul.autumnboot.rest.common.logging;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * 警告日志计数器
@@ -15,12 +15,12 @@ public final class WarningCounter {
      * {@link LoggerInstrument}向{@link ch.qos.logback.classic.Logger#warn(String)}等写入警告日志的方法
      * 插入了调用{@link WarnCounter#increment()}的逻辑
      */
-    private static final AtomicLong WARNINGS = new AtomicLong(0);
+    private static final AtomicInteger WARNINGS = new AtomicInteger(0);
 
     /**
      * 获取通过{@link ch.qos.logback.classic.Logger}打印的警告日志数量
      */
-    public static long getWarnings() {
+    public static int getWarnings() {
         synchronized (WARNINGS) {
             return WARNINGS.get();
         }
